+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 2.3.1 (...)
 - static IP can also not be network and broadcast of `poolRange`
+- do not allow to set static IP addresses already in use
 
 ## 2.3.0 (2016-02-19)
 - delete routes without specifying `dev`